requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/test.txt requirements/base.in requirements/test.in
#
coverage==4.2             # via pytest-cov
ddt==1.1.0
djangorestframework==3.4.7
freezegun==0.3.7
funcsigs==1.0.2           # via mock
mock==2.0.0
pbr==1.10.0               # via mock
py==1.4.31                # via pytest, pytest-catchlog
pytest-catchlog==1.2.2
pytest-cov==2.3.1
pytest-django==3.0.0
pytest==3.0.3             # via pytest-catchlog, pytest-cov, pytest-django
python-dateutil==2.5.3    # via freezegun
six==1.10.0               # via freezegun, mock, python-dateutil
